requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.69
colorama==0.4.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.12.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
outcome==1.2.0
psycopg2==2.9.5
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-daemon==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
schedule==1.1.0
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.7
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
yarl==1.8.2
zipp==3.12.0
